backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
pydantic-settings>=2.0
alembic>=1.14
httpx>=0.28.0
python-multipart>=0.0.12
cryptography>=44.0
# Azure Key Vault
azure-identity>=1.19.0
azure-keyvault-secrets>=4.9.0
# Keycloak OIDC auth
python-jose[cryptography]>=3.3.0
